Added tests and made custom 404 handler

+++ rest_server/server.py
@@ -12,7 +12,7 @@ Functions:
 
 '''
 
-from flask import Flask, request, redirect, jsonify, url_for
+from flask import Flask, request, redirect, abort, jsonify, url_for
 
 from .db import init as db_init
 from .db import db, Shoes, shoe_schema, shoes_schema
@@ -29,11 +29,17 @@ def init(db_user: str, db_pass: str, db_host: str, db_name: str):
     app = db_init(app, db_user, db_pass, db_host, db_name)
 
 
-# Endpoints
 @app.route('/')
 def index():
     return redirect(url_for('get_shoes'), code=302)
 
+
+@app.errorhandler(404)
+def shoe_not_found(e):
+    return '<h1>404 Not Found<h1>\n<p>Resource you requested does not exist.</p>', 404
+
+
+# Endpoints
 
 # Create shoe
 @app.route('/shoe', methods=['POST'])
@@ -63,6 +69,8 @@ def get_shoes():
 @app.route('/shoe/<id>', methods=['GET'])
 def get_shoe(id):
     shoe = Shoes.query.get(id)
+    if shoe is None:
+        abort(404)
     return shoe_schema.jsonify(shoe)
 
 
@@ -70,7 +78,9 @@ def get_shoe(id):
 @app.route('/shoe/<id>', methods=['PUT'])
 def update_shoe(id):
     shoe = Shoes.query.get(id)
-    
+    if shoe is None:
+        abort(404)
+
     shoe.size         = int(request.json['size'])    if 'size' in request.json else shoe.size
     shoe.shoes_type   = request.json['shoes_type']   if 'shoes_type' in request.json else shoe.shoes_type
     shoe.manufacturer = request.json['manufacturer'] if 'manufacturer' in request.json else shoe.manufacturer
@@ -84,6 +94,8 @@ def update_shoe(id):
 @app.route('/shoe/<id>', methods=['DELETE'])
 def delete_shoe(id):
     shoe = Shoes.query.get(id)
+    if shoe is None:
+        abort(404)
     db.session.delete(shoe)
     db.session.commit()
